requirements.txt
dnspython==2.8.0
msgspec==0.18.6
pymongo==4.8.0
python-dotenv==1.0.1
redis==5.0.8
